refactor(http): introduce dcc-mcp-http-types crate (#852)
First landing zone for the dcc-mcp-http Clean-Architecture split called
for in #852. dcc-mcp-http-types hosts the wire-level value types — pure
structs with no axum, reqwest, tokio, or pyo3 dependency. The dependency
direction is strictly inward:

    dcc-mcp-http (server + pyo3 bindings)  →  dcc-mcp-http-types (types)

Migrated this round:
- TruncationEnvelope     (issue #771 truncation metadata envelope)
- SseChunkFrame          (issue #771 SSE chunking wire frame)
- chunk_sse_data()       (payload → frames helper)
- format_chunked_sse()   (frames → SSE text events helper)

Compatibility:
- crates/dcc-mcp-http/src/payload.rs now pub-use re-exports these
  symbols under their historical paths, so the existing pub use in
  dcc-mcp-http::lib.rs keeps the public API stable for downstream
  Python and Rust consumers.

Why these types as the seed:
- Smallest, most self-contained wire types in dcc-mcp-http.
- Only serde + base64 + serde_json as third-party deps — no gateway,
  no transport, no pyo3 — so the dependency direction is trivial to
  verify.
- Ships #[must_use] on the two free-functions and TruncationEnvelope::new
  so callers cannot silently drop chunking / truncation results (contract
  spirit).

Tests: 12 unit tests in dcc-mcp-http-types (migrated from payload.rs)
plus 230 in dcc-mcp-http all pass. cargo clippy -p dcc-mcp-http-types
--all-targets -- -D warnings is clean; cargo fmt --all -- --check is
clean.

Follow-up PRs in the #852 chain will migrate HttpError, JobRecoveryPolicy,
MinimalModeConfig, and ultimately carve up McpHttpConfig along the same
lines as DccServerOptions (gateway / observability / execution /
diagnostics), keeping Rust and Python configs in lockstep.

Refs: #852, pairs with #845 (gateway split). Part of #848 epic.
