+++ crates/dcc-mcp-http-types/Cargo.toml
@@ -1,0 +1,20 @@
+[package]
+name = "dcc-mcp-http-types"
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+authors.workspace = true
+license.workspace = true
+repository.workspace = true
+readme = "README.md"
+description = "Wire-level value types for the DCC MCP HTTP server — no axum/tokio/reqwest."
+keywords = ["mcp", "http", "types", "clean-architecture"]
+categories = ["api-bindings"]
+
+[lints]
+workspace = true
+
+[dependencies]
+serde = { workspace = true }
+base64 = { workspace = true }
+serde_json = { workspace = true }
