+++ crates/dcc-mcp-http/Cargo.toml
@@ -10,6 +10,7 @@ description = "MCP Streamable HTTP server for DCC applications (axum-based, thre
 
 [dependencies]
 # Internal crates
+dcc-mcp-http-types = { path = "../dcc-mcp-http-types" }
 dcc-mcp-actions = { path = "../dcc-mcp-actions" }
 dcc-mcp-naming = { path = "../dcc-mcp-naming" }
 dcc-mcp-protocols = { path = "../dcc-mcp-protocols" }
